+++ CHANGES.md
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+0.2.0 (UNRELEASED)
+------------------
+
+### Changed
+
+- One client/connection per IRC network or Matrix server.
+- Channels in the same network/server are joined by the same client.
+
+
 0.1.0 (2022-12-28)
 ------------------
 
@@ -8,3 +17,4 @@ Changelog
 
 - Initial release.
 - Bridging between multiple IRC channels and Matrix rooms.
+- One client/connection per channel/room.